Eliminating `Birecursive`.
`Birecursive` is misleading because in (almost) all cases, one of the super
class instances is total and the other is not.

It previously existed because the `Steppable` functions were split between
`Recursive` and `Corecursive`, so it often _appeared_ that a `Birecursive`
instance was necessary, since many operations require either `project` or
`embed`, but rarely are both `cata` and `ana` required. Splitting out
`Steppable` makes it possible to eliminate most places that had both `Recursive`
and `Corecursive` constraints.`